Some changes to clean up the code
