+++ DuckDuckGo/NetworkProtection/AppTargets/BothAppTargets/VPNUIActionHandler.swift
@@ -21,8 +21,6 @@ import Foundation
 import NetworkProtectionIPC
 import NetworkProtectionProxy
 import NetworkProtectionUI
-import PixelKit
-import VPNPixels
 import VPNAppLauncher
 
 /// Main App's VPN UI action handler
@@ -31,18 +29,15 @@ final class VPNUIActionHandler: VPNUIActionHandling {
 
     private let vpnIPCClient: VPNControllerXPCClient
     private let proxySettings: TransparentProxySettings
-    private let pixelKit: PixelFiring?
     private let vpnURLEventHandler: VPNURLEventHandler
 
     init(vpnIPCClient: VPNControllerXPCClient = .shared,
          vpnURLEventHandler: VPNURLEventHandler,
-         proxySettings: TransparentProxySettings,
-         pixelKit: PixelFiring? = PixelKit.shared) {
+         proxySettings: TransparentProxySettings) {
 
         self.vpnIPCClient = vpnIPCClient
         self.vpnURLEventHandler = vpnURLEventHandler
         self.proxySettings = proxySettings
-        self.pixelKit = pixelKit
     }
 
     public func moveAppToApplications() async {
@@ -52,10 +47,6 @@ final class VPNUIActionHandler: VPNUIActionHandling {
     }
 
     func setExclusion(_ exclude: Bool, forDomain domain: String) async {
-
-        let engagementPixel = DomainExclusionsEngagement(exclude: exclude, domain: domain)
-        pixelKit?.fire(engagementPixel)
-
         proxySettings.setExclusion(exclude, forDomain: domain)
         try? await vpnIPCClient.command(.restartAdapter)
     }
