+++ LocalPackages/NetworkProtectionMac/README.md
@@ -4,3 +4,4 @@ This package contains several components of VPN that are specific to macOS.
 
 - NetworkProtectionUI: UI elements used in the VPN that are reusable across build targets.
 - NetworkProtectionIPC: code to support IPC communication between VPN targets.
+- VPNPixels: module for NetworkProtectionUI to have pixel support for shared UI components.
